dataframe rows seriation, which will reorder the rows in a better pattern

dataframe rows seriation, which will reorder the rows in a better pattern

Usage

seriate_df(x)

Arguments

x

dataframe

Value

seriated dataframe

Examples

x <- mini_diamond %>%
  dplyr::select(id, dplyr::where(is.numeric)) %>%
  dplyr::mutate(
    dplyr::across(
      dplyr::where(is.numeric),
      ~ round(.x / max(.x), 4)
    )
  ) %>%
  c2r("id")

seriate_df(x)
#>         carat  price      x      y
#> id-7   0.1034 0.0417 0.4734 0.4749
#> id-15  0.1073 0.0340 0.4723 0.4807
#> id-94  0.1073 0.0256 0.4838 0.4936
#> id-60  0.1111 0.0337 0.4931 0.5006
#> id-61  0.1149 0.0350 0.4885 0.5018
#> id-42  0.1226 0.0327 0.5000 0.5076
#> id-88  0.1188 0.0348 0.5058 0.5064
#> id-87  0.1303 0.0331 0.5081 0.5181
#> id-26  0.1188 0.0398 0.5035 0.5146
#> id-86  0.1188 0.0494 0.5058 0.5134
#> id-14  0.1149 0.0450 0.5000 0.5123
#> id-39  0.1149 0.0479 0.5000 0.5088
#> id-57  0.1149 0.0479 0.4988 0.5064
#> id-89  0.1149 0.0670 0.5162 0.5076
#> id-96  0.1226 0.0474 0.5139 0.5204
#> id-53  0.1264 0.0584 0.5277 0.5309
#> id-69  0.1533 0.0272 0.5358 0.5461
#> id-16  0.1571 0.0259 0.5427 0.5531
#> id-54  0.1533 0.0516 0.5450 0.5543
#> id-81  0.1571 0.0529 0.5508 0.5589
#> id-72  0.1609 0.0578 0.5450 0.5578
#> id-21  0.1533 0.0599 0.5439 0.5519
#> id-73  0.1533 0.0682 0.5462 0.5554
#> id-56  0.1533 0.0621 0.5485 0.5601
#> id-18  0.1303 0.0561 0.5543 0.5554
#> id-46  0.1303 0.0577 0.5450 0.5566
#> id-95  0.1418 0.0799 0.5404 0.5519
#> id-74  0.1724 0.0859 0.5600 0.5578
#> id-37  0.1954 0.1560 0.5947 0.5963
#> id-45  0.1801 0.1227 0.5878 0.5811
#> id-3   0.1992 0.1126 0.5947 0.6044
#> id-62  0.1762 0.1002 0.5912 0.6044
#> id-84  0.1916 0.0785 0.5901 0.5893
#> id-8   0.1954 0.0571 0.5831 0.5928
#> id-30  0.1992 0.0777 0.6074 0.6068
#> id-13  0.2146 0.0906 0.6132 0.6208
#> id-67  0.2222 0.0800 0.6236 0.6254
#> id-19  0.2261 0.1195 0.6166 0.6289
#> id-65  0.2414 0.1083 0.6189 0.6313
#> id-76  0.2452 0.0961 0.6524 0.6289
#> id-10  0.2682 0.0938 0.6420 0.6313
#> id-23  0.2682 0.0642 0.6513 0.6418
#> id-75  0.2720 0.1229 0.6490 0.6523
#> id-52  0.2759 0.1279 0.6536 0.6663
#> id-29  0.2759 0.1276 0.6605 0.6744
#> id-5   0.2759 0.1386 0.6617 0.6733
#> id-77  0.2720 0.1455 0.6732 0.6779
#> id-55  0.2720 0.1699 0.6547 0.6499
#> id-44  0.2682 0.1713 0.6339 0.6488
#> id-78  0.2682 0.1961 0.6570 0.6674
#> id-83  0.2797 0.1934 0.6663 0.6791
#> id-12  0.2720 0.1778 0.6778 0.6779
#> id-80  0.3065 0.2258 0.6824 0.6954
#> id-24  0.3525 0.2356 0.7055 0.7200
#> id-70  0.3448 0.2373 0.7229 0.7340
#> id-59  0.3487 0.2283 0.7367 0.7468
#> id-41  0.3755 0.2149 0.7333 0.7456
#> id-50  0.3831 0.2691 0.7240 0.7363
#> id-64  0.3870 0.2485 0.7321 0.7340
#> id-98  0.3831 0.2392 0.7240 0.7270
#> id-1   0.3908 0.1679 0.7217 0.7211
#> id-58  0.3448 0.1824 0.7044 0.7141
#> id-22  0.3448 0.1801 0.7113 0.7083
#> id-66  0.3448 0.1562 0.7021 0.7048
#> id-82  0.3448 0.1189 0.7032 0.7060
#> id-71  0.3716 0.1315 0.7321 0.7328
#> id-17  0.3716 0.1242 0.7390 0.7503
#> id-34  0.4368 0.1291 0.7656 0.7643
#> id-20  0.4598 0.1670 0.7633 0.7631
#> id-90  0.4330 0.2040 0.7679 0.7806
#> id-9   0.3870 0.3101 0.7425 0.7468
#> id-36  0.3831 0.3392 0.7229 0.7246
#> id-35  0.3831 0.3625 0.7356 0.7480
#> id-51  0.3831 0.3743 0.7298 0.7351
#> id-11  0.3908 0.4361 0.7356 0.7468
#> id-31  0.3946 0.4659 0.7552 0.7585
#> id-85  0.3831 0.3929 0.7818 0.7830
#> id-100 0.4598 0.3519 0.7760 0.7795
#> id-33  0.4636 0.2914 0.7656 0.7830
#> id-43  0.4291 0.3044 0.7483 0.7608
#> id-93  0.4176 0.2982 0.7621 0.7748
#> id-38  0.4176 0.3007 0.7644 0.7783
#> id-47  0.4215 0.2979 0.7691 0.7818
#> id-79  0.4138 0.2869 0.7968 0.7935
#> id-27  0.4330 0.2773 0.8002 0.8028
#> id-63  0.5172 0.3120 0.8060 0.8086
#> id-40  0.5824 0.4099 0.8349 0.8390
#> id-4   0.5900 0.5244 0.8580 0.8693
#> id-91  0.6628 0.4643 0.8776 0.8821
#> id-32  0.7663 0.3144 0.8984 0.9032
#> id-28  0.7739 0.3520 0.9088 0.9102
#> id-48  0.7701 0.4046 0.9584 0.9557
#> id-49  0.8276 0.4831 0.9596 0.9638
#> id-97  1.0000 0.7647 1.0000 1.0000
#> id-68  0.8889 1.0000 0.9781 0.9697
#> id-6   0.7739 0.7811 0.9619 0.9767
#> id-25  0.8046 0.8780 0.9203 0.9242
#> id-99  0.7663 0.8516 0.8811 0.8856
#> id-92  0.5785 0.8129 0.8291 0.8448
#> id-2   0.5785 0.6516 0.8395 0.8378